requirements.txt
requests
pandas
tabula-py
matplotlib
bs4
flake8 
pytest
prophet